Issue #154 refactor: repository 패키지 구조 리팩토링
