--- app/src/main/java/com/lateinit/rightweight/data/repository/BaseRepository.kt
@@ -1,4 +1,0 @@
-package com.lateinit.rightweight.data.repository
-
-class BaseRepository {
-}
