+++ app/src/main/java/com/lateinit/rightweight/data/repository/impl/UserRepositoryImpl.kt
@@ -1,10 +1,11 @@
-package com.lateinit.rightweight.data.repository
+package com.lateinit.rightweight.data.repository.impl
 
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
 import com.lateinit.rightweight.data.datasource.UserLocalDataSource
 import com.lateinit.rightweight.data.datasource.UserRemoteDataSource
 import com.lateinit.rightweight.data.model.User
 import com.lateinit.rightweight.data.model.WriteModelData
+import com.lateinit.rightweight.data.repository.UserRepository
 import com.lateinit.rightweight.ui.model.HistoryUiModel
 import com.lateinit.rightweight.util.toHistoryUiModel
 import kotlinx.coroutines.flow.Flow
